move blue-green flag to role
