+++ plugins/kubernetes/app/views/kubernetes/roles/_form.html.erb
@@ -11,6 +11,7 @@
           help: "Name for the resource this role will create (service/job), a-Z 0-9 and -"
     %>
     <%= form.input :autoscaled, label: "Replicas managed externally", as: :check_box, help: "Deploying will not change replicas count. Use this when autoscaling deployments." %>
+    <%= form.input :blue_green, as: :check_box, label: "Blue/Green Deployment", help: "Starts a new isolated deployment shifting between blue and green sufixes, switching service selectors if successfully deployed and deleting previous resources, all active resources must be deleted manually when switching this" %>
 
     <div class="form-group">
       <div class="col-lg-offset-2 col-lg-10">
