--- plugins/kubernetes/db/migrate/20180112113329_add_blue_green_to_stages.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-class AddBlueGreenToStages < ActiveRecord::Migration[5.1]
-  def change
-    add_column :stages, :blue_green, :boolean, default: false, null: false
-  end
-end
