--- plugins/kubernetes/db/migrate/20180112113400_add_blue_phase_to_kubernetes_releases.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-class AddBluePhaseToKubernetesReleases < ActiveRecord::Migration[5.1]
-  def change
-    add_column :kubernetes_releases, :blue_phase, :boolean, default: false, null: false
-  end
-end
